Prepare platform for 386 merge; merge 386_39179 + binary blobs (RT-AC86U)
- Reorganized all prebuild_arm_* folders to use model-based paths
  (like rc prebuild).  Symlinks are used for common files.
  This is necessary as the use of the EX7 and EXHND variables was
  non-consistent between mainline and ax, causing conflicts.
  Makefiles will now always use prebuild/$(BUILD_NAME), with
  symlinks handling the rest.
- Merged GPL 386_39179 + binary blobs from RT-AC86U
- Updated rc/openvpn.c and httpd for OpenSSL 1.1 compatibility
- Removed kludges for 382 models
- Rewritten networkmap traffic meters for new style
- Harmonized OpenSSL and inadyn code with upstream
- Re-synced rstats code with upstream (hopefully the
  random rate spikes are fixed) since it carries
  many important enhancements for LACP and others.

+++ release/src/router/mDNSResponder/mDNSVxWorks/mDNSVxWorks.c
@@ -905,7 +905,7 @@ mDNSlocal mStatus	UpdateInterfaceList( mDNS *const inMDNS, mDNSs32 inUTC )
 			mDNSPlatformMemZero( &ifr6, sizeof( ifr6 ) );
 			strcpy( ifr6.ifr_name, ifa->ifa_name );
 			ifr6.ifr_addr = *sa6;
-			if( ioctl( infoSock, SIOCGIFAFLAG_IN6, (int) &ifr6 ) != -1 )
+			if( ioctl( infoSock, SIOCGIFAFLAG_IN6, (long) &ifr6 ) != -1 )
 			{
 				flags = ifr6.ifr_ifru.ifru_flags6;
 			}
@@ -1456,7 +1456,7 @@ mDNSlocal mStatus	SetupSocket( mDNS *const inMDNS, const mDNSAddr *inAddr, mDNSB
 	
 	// Make the socket non-blocking so we can potentially get multiple packets per select call.
 	
-	err = ioctl( sock, FIONBIO, (int) &on );
+	err = ioctl( sock, FIONBIO, (long) &on );
 	check_translated_errno( err == 0, errno_compat(), kOptionErr );
 	
 	*sockPtr = sock;
